[frontend] Root with ts, Routes created to mount component  (#10170)
